Fix compat with xp through 7
WriteFile requires a non NULL 4th parameter, this requirement was unoffically removed on windows 8 and up which is loader and packer worked on newer oses.
CreateProcess appears to only error with ERROR_EXE_MACHINE_TYPE_MISMATCH if windows is aware of the machine type the binary is for which isn't the case for x86_64 on windows xp x86_32 so it just gives ERROR_BAD_EXE_FORMAT. Not sure if this is the case for aarch32 and aarch64 binaries on any other versions of windows pre 8.0 as well.
Calling AttachConsole on at least xp in a console subsystem program prevents accessing the calling console if one exists. I expect that this is the intended behaviour as it is meant be wrapped in a if block with !AttachConsole on such programs anyway. Loader didn't need this call anymore so I just removed it.