Deduplicate Sample project files
May end up finding a way to merge the projects while retaining 4 separate builds without using batch build as well but this is hopefully slightly better for now.
Next up is repeating this for the other projects with top level versions of the msbuild files as most settings are duplicated the other projects as well.
This was mostly done to reduce the amount of time I spent messing with the project files for all of the projects when changing subsystems and what not.
At some point I plan to create actual manifest files as well that will be reused between projects and so these files will be a great place to put them.