Use central properties for Fatpack GUI
This is definitely not an intended use for msbuild property sheets, a requirement for it to work at all is to have each prop file imported before any related properties are defined. i.e. include the configuration sheet before using any configuration PropertyGroups, otherwise the property editor in visual studio doesn't know the value of most properties and either leaves it black or uses the default value. In spite of that, the project builds without issue and with the imports in the correct places the property editor does show everything.