Restore xp support
Had to disable the uac manifest entry for anything to run though. This is weird as it could be there if using msvc instead of clang.
Fixed error reporting when CreateFile fails.
Disable emitting debug information in the release build for a small size reduction. Might revert if having the debug info is worth it.